Fetch: separated ngx_js_http_* from ngx_js_fetch_*.
